requirements.txt
allure-pytest==2.8.37
selenium==4.1.0
pytest==6.2.5